+++ docs/Changelog.md
@@ -1,7 +1,7 @@
 
 # Changelog
 
-1.8.1 (not released yet)
+1.8.1 (28 Feb 2023)
 ------------------------
 * [#4018](https://github.com/TouK/nussknacker/pull/4018) Fix for: generate test data mechanism didn't work for json messages with defined schema id 
 * [#4024](https://github.com/TouK/nussknacker/pull/4024) Fix encoding object in sink with JSON schema pattern properties